// public/static/charting_library/charting_library.d.ts
export declare type Nominal<T, Name extends string> = T & {
	[Symbol.species]: Name;
};
// ...
export interface ChartData {
	/** unique ID of the chart (may be `undefined` if it wasn't saved before) */
	id: string | undefined;
	/** name of the chart */
	name: string;
	/** symbol of the chart */
	symbol: string;
	/** resolution of the chart */
	resolution: ResolutionString;
	/** content of the chart */
	content: string;
}
/**
 * Meta information about a saved chart
 */
export interface ChartMetaInfo {
	/** unique ID of the chart. */
	id: number;
	/** name of the chart */
	name: string;
	/** symbol of the chart */
	symbol: string;
	/** resolution of the chart */
	resolution: ResolutionString;
	/** UNIX time when the chart was last modified */
	timestamp: number;
}
// ...
export interface ChartTemplate {
	/**
	 * The template content.
	 */
	content?: ChartTemplateContent;
}
/**
 * Chart template content. The properties of the chart that are saved/loaded when the library saves/loads a chart template.
 */
export interface ChartTemplateContent {
	[key: string]: any;
	/**
	 * Chart properties (for example color, etc).
	 */
	chartProperties?: {
		/**
		 * Chart pane properties.
		 */
		paneProperties: any;
		/**
		 * Chart scales properties.
		 */
		scalesProperties: any;
	};
	/**
	 * Series properties (for example chart style, etc).
	 */
	mainSourceProperties?: any;
	/**
	 * The version of the chart template.
	 */
	version?: number;
}
// ...
export interface IExternalSaveLoadAdapter {
	/**
	 * Get all saved charts.
	 * @returns Array of chart meta information
	 */
	getAllCharts(): Promise<ChartMetaInfo[]>;
	/**
	 * Remove a chart.
	 * @param  id - Unique ID of the chart (see {@link getAllCharts})
	 */
	removeChart<T extends number | string>(id: T): Promise<void>;
	/**
	 * Save the chart
	 * @param  {ChartData} chartData - Chart description data
	 * @returns unique ID of the chart
	 */
	saveChart(chartData: ChartData): Promise<string>;
	/**
	 * Load the chart from the server
	 * @param  {number} chartId - Unique ID of the chart to load (see {@link getAllCharts})
	 * @returns chart content contained in the `content` field when saving the chart ({@link ChartData})
	 */
	getChartContent(chartId: number): Promise<string>;
	/**
	 * Get all saved study templates
	 * @returns Array of study template meta information
	 */
	getAllStudyTemplates(): Promise<StudyTemplateMetaInfo[]>;
	/**
	 * Remove a study template
	 * @param  {StudyTemplateMetaInfo} studyTemplateInfo
	 */
	removeStudyTemplate(studyTemplateInfo: StudyTemplateMetaInfo): Promise<void>;
	/**
	 * Save a study template
	 * @param  {StudyTemplateData} studyTemplateData - Study template data to save
	 */
	saveStudyTemplate(studyTemplateData: StudyTemplateData): Promise<void>;
	/**
	 * load a study template from the server
	 * @param  {StudyTemplateMetaInfo} studyTemplateInfo
	 * @returns Study template `content`
	 */
	getStudyTemplateContent(studyTemplateInfo: StudyTemplateMetaInfo): Promise<string>;
	/**
	 * Get names of all saved drawing templates
	 * @param  {string} toolName - name of the drawing tool
	 * @returns names of saved drawing templates
	 */
	getDrawingTemplates(toolName: string): Promise<string[]>;
	/**
	 * Load a drawing template from the server
	 * @param  {string} toolName - name of the drawing tool
	 * @param  {string} templateName - name of the template
	 * @returns content of the drawing template
	 */
	loadDrawingTemplate(toolName: string, templateName: string): Promise<string>;
	/**
	 * Remove a drawing template
	 * @param  {string} toolName - name of the drawing tool
	 * @param  {string} templateName - name of the template
	 */
	removeDrawingTemplate(toolName: string, templateName: string): Promise<void>;
	/**
	 * Save a drawing template
	 * @param  {string} toolName - name of the drawing tool
	 * @param  {string} templateName - name of the template
	 * @param  {string} content - content of the drawing template
	 */
	saveDrawingTemplate(toolName: string, templateName: string, content: string): Promise<void>;
	/**
	 * Load a chart template from the server
	 *
	 * @param templateName The name of the template.
	 *
	 * @returns The chart template content.
	 */
	getChartTemplateContent(templateName: string): Promise<ChartTemplate>;
	/**
	 * Get names of all saved chart templates.
	 *
	 * @returns An array of names.
	 */
	getAllChartTemplates(): Promise<string[]>;
	/**
	 * Save a chart template.
	 *
	 * @param newName The name of the template.
	 * @param theme The template content.
	 */
	saveChartTemplate(newName: string, theme: ChartTemplateContent): Promise<void>;
	/**
	 * Remove a chart template.
	 *
	 * @param templateName The name of the template.
	 */
	removeChartTemplate(templateName: string): Promise<void>;
}
// ...
export interface StudyTemplateData {
	/**
	 * Template name.
	 */
	name: string;
	/**
	 * Template content.
	 */
	content: string;
}
/**
 * Study template metainfo.
 */
export interface StudyTemplateMetaInfo {
	/**
	 * Template name.
	 */
	name: string;
}
// ...
export type ResolutionString = Nominal<string, "ResolutionString">;